refactor: inline function to meet 5 function requirement

+++ pipex.h
@@ -6,7 +6,7 @@
 /*   By: dmontesd <dmontesd@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/27 19:06:51 by dmontesd          #+#    #+#             */
-/*   Updated: 2025/04/07 05:22:42 by dmontesd         ###   ########.fr       */
+/*   Updated: 2025/04/07 05:33:55 by dmontesd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef PIPEX_H
@@ -123,7 +123,7 @@ void	path_iter_free(t_path_iter *iter);
  */
 typedef struct s_command
 {
-	int			argc;
+	size_t		argc;
 	char		**args;
 	char		*redirection;
 	int			redirect_fd;
